Move scoreboard functions to constant
